HP-182 Remove hard dependency to drf-oidc-auth
The drf-oidc-auth brings in the whole Django REST Framework (DFR).
That's a bit too much for projects that don't otherwise use DRF.

Users who want to still use the helusers.oidc.ApiTokenAuthentication
class need to make sure that they install the drf-oidc-auth dependency
by themselves.
