+++ requirements-dev.in
@@ -1,3 +1,4 @@
 django-allauth
+drf-oidc-auth
 pytest
 pytest-django
