+++ setup.py
@@ -22,7 +22,6 @@ setup(
     author_email='dev@hel.fi',
     install_requires=[
         'Django',
-        'drf-oidc-auth>=0.9',
         'requests',
         'python-jose',
     ],
